requirements.txt
pandas
numpy
geopy
streamlit
plotly.express
statsmodels